Refactor: move formatters dedicated to querying objects to Formatters::Query
